+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - `merge`: merge search results from multiple databases.
 - `compute`:
     - fix splitting very short genomes.
+    - remove flag `-e/--exact-number`, making it default.
 - `index`:
     - do not roundup sizes of indexes. The searching speed is not
       affected and even faster due to optimization of `search` command.